requirements.txt
python-telegram-bot==20.3
tradingview-ta==3.3.0
pytz==2023.3
